requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
environs==11.0.0
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi_cors==0.0.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
marshmallow==3.23.1
mysql-connector-python==9.1.0
numpy==2.1.3
packaging==24.1
pandas==2.2.3
playwright==1.48.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
